+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 5.1.2
+
+* Fix: Servlet 3.1 compatibility issue (#1681)
+* Fix: Do not drop Contexts key if Collection, Array or Char (#1680)
+
 ## 5.1.1
 
 * Fix: Remove onActivityPreCreated call in favor of onActivityCreated (#1661)